requirements.txt
tensorflow-gpu
numpy
apache_beam
PyYAML
faiss-cpu
jupyter
matplotlib
httplib2==0.19.0
tensorflow_probability
transformers==4.17.0
einops
tinycss
google-cloud-storage
hydra-core
pandas
selenium
fsspec
dacite
skia-python